simplify IncrementalIndex since group-by v1 has been removed (#15448)
