+++ benchmarks/src/test/java/org/apache/druid/benchmark/query/GroupByBenchmark.java
@@ -608,7 +608,6 @@ public class GroupByBenchmark
                 .withRollup(withRollup)
                 .build()
         )
-        .setConcurrentEventAdd(true)
         .setMaxRowCount(rowsPerSegment)
         .build();
   }
