+++ processing/src/main/java/org/apache/druid/segment/incremental/OnheapIncrementalIndex.java
@@ -47,14 +47,23 @@ import org.apache.druid.utils.JvmUtils;
 import javax.annotation.Nullable;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Comparator;
+import java.util.Deque;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentLinkedDeque;
+import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.ConcurrentNavigableMap;
+import java.util.concurrent.ConcurrentSkipListMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.Function;
+import java.util.stream.Stream;
 
 /**
  *
@@ -72,7 +81,7 @@ public class OnheapIncrementalIndex extends IncrementalIndex
   private static final long ROLLUP_RATIO_FOR_AGGREGATOR_FOOTPRINT_ESTIMATION = 100;
 
   /**
-   * overhead per {@link ConcurrentHashMap.Node}  or {@link java.util.concurrent.ConcurrentSkipListMap.Node} object
+   * overhead per {@link ConcurrentSkipListMap.Node} object in facts table
    */
   private static final int ROUGH_OVERHEAD_PER_MAP_ENTRY = Long.BYTES * 5 + Integer.BYTES;
   private final ConcurrentHashMap<Integer, Aggregator[]> aggregators = new ConcurrentHashMap<>();
@@ -118,28 +127,19 @@ public class OnheapIncrementalIndex extends IncrementalIndex
 
   OnheapIncrementalIndex(
       IncrementalIndexSchema incrementalIndexSchema,
-      boolean deserializeComplexMetrics,
-      boolean concurrentEventAdd,
-      boolean sortFacts,
       int maxRowCount,
       long maxBytesInMemory,
-      // preserveExistingMetrics should only be set true for DruidInputSource since that is the only case where we can have existing metrics
-      // This is currently only use by auto compaction and should not be use for anything else.
+      // preserveExistingMetrics should only be set true for DruidInputSource since that is the only case where we can
+      // have existing metrics. This is currently only use by auto compaction and should not be use for anything else.
       boolean preserveExistingMetrics,
       boolean useMaxMemoryEstimates
   )
   {
-    super(
-        incrementalIndexSchema,
-        deserializeComplexMetrics,
-        concurrentEventAdd,
-        preserveExistingMetrics,
-        useMaxMemoryEstimates
-    );
+    super(incrementalIndexSchema, preserveExistingMetrics, useMaxMemoryEstimates);
     this.maxRowCount = maxRowCount;
     this.maxBytesInMemory = maxBytesInMemory == 0 ? Long.MAX_VALUE : maxBytesInMemory;
-    this.facts = incrementalIndexSchema.isRollup() ? new RollupFactsHolder(sortFacts, dimsComparator(), getDimensions())
-                                                   : new PlainFactsHolder(sortFacts, dimsComparator());
+    this.facts = incrementalIndexSchema.isRollup() ? new RollupFactsHolder(dimsComparator(), getDimensions())
+                                                   : new PlainFactsHolder(dimsComparator());
     maxBytesPerRowForAggregators =
         useMaxMemoryEstimates ? getMaxBytesPerRowForAggregators(incrementalIndexSchema) : 0;
     this.useMaxMemoryEstimates = useMaxMemoryEstimates;
@@ -190,9 +190,7 @@ public class OnheapIncrementalIndex extends IncrementalIndex
   @Override
   protected void initAggs(
       final AggregatorFactory[] metrics,
-      final Supplier<InputRow> rowSupplier,
-      final boolean deserializeComplexMetrics,
-      final boolean concurrentEventAdd
+      final Supplier<InputRow> rowSupplier
   )
   {
     selectors = new HashMap<>();
@@ -200,18 +198,14 @@ public class OnheapIncrementalIndex extends IncrementalIndex
     for (AggregatorFactory agg : metrics) {
       selectors.put(
           agg.getName(),
-          new CachingColumnSelectorFactory(
-              makeColumnSelectorFactory(agg, rowSupplier, deserializeComplexMetrics),
-              concurrentEventAdd
-          )
+          new CachingColumnSelectorFactory(makeColumnSelectorFactory(agg, rowSupplier))
       );
       if (preserveExistingMetrics) {
         AggregatorFactory combiningAgg = agg.getCombiningFactory();
         combiningAggSelectors.put(
             combiningAgg.getName(),
             new CachingColumnSelectorFactory(
-                makeColumnSelectorFactory(combiningAgg, rowSupplier, deserializeComplexMetrics),
-                concurrentEventAdd
+                makeColumnSelectorFactory(combiningAgg, rowSupplier)
             )
         );
       }
@@ -550,6 +544,7 @@ public class OnheapIncrementalIndex extends IncrementalIndex
    * If preserveExistingMetrics flag is set, then this method will combine values from two aggregators, the aggregator
    * for aggregating from input into output field and the aggregator for combining already aggregated field, as needed
    */
+  @Nullable
   private <T> Object getMetricHelper(AggregatorFactory[] metrics, Aggregator[] aggs, int aggOffset, Function<Aggregator, T> getMetricTypeFunction)
   {
     if (preserveExistingMetrics) {
@@ -605,18 +600,13 @@ public class OnheapIncrementalIndex extends IncrementalIndex
    */
   static class CachingColumnSelectorFactory implements ColumnSelectorFactory
   {
-    private final Map<String, ColumnValueSelector<?>> columnSelectorMap;
+    private final HashMap<String, ColumnValueSelector<?>> columnSelectorMap;
     private final ColumnSelectorFactory delegate;
 
-    public CachingColumnSelectorFactory(ColumnSelectorFactory delegate, boolean concurrentEventAdd)
+    public CachingColumnSelectorFactory(ColumnSelectorFactory delegate)
     {
       this.delegate = delegate;
-
-      if (concurrentEventAdd) {
-        columnSelectorMap = new ConcurrentHashMap<>();
-      } else {
-        columnSelectorMap = new HashMap<>();
-      }
+      this.columnSelectorMap = new HashMap<>();
     }
 
     @Override
@@ -628,7 +618,7 @@ public class OnheapIncrementalIndex extends IncrementalIndex
     @Override
     public ColumnValueSelector<?> makeColumnValueSelector(String columnName)
     {
-      ColumnValueSelector existing = columnSelectorMap.get(columnName);
+      ColumnValueSelector<?> existing = columnSelectorMap.get(columnName);
       if (existing != null) {
         return existing;
       }
@@ -656,9 +646,6 @@ public class OnheapIncrementalIndex extends IncrementalIndex
     {
       return new OnheapIncrementalIndex(
           Objects.requireNonNull(incrementalIndexSchema, "incrementIndexSchema is null"),
-          deserializeComplexMetrics,
-          concurrentEventAdd,
-          sortFacts,
           maxRowCount,
           maxBytesInMemory,
           preserveExistingMetrics,
@@ -732,6 +719,196 @@ public class OnheapIncrementalIndex extends IncrementalIndex
     public int hashCode()
     {
       return Objects.hash(preserveExistingMetrics);
+    }
+  }
+
+  static final class RollupFactsHolder implements FactsHolder
+  {
+    // Can't use Set because we need to be able to get from collection
+    private final ConcurrentNavigableMap<IncrementalIndexRow, IncrementalIndexRow> facts;
+    private final List<DimensionDesc> dimensionDescsList;
+
+    RollupFactsHolder(
+        Comparator<IncrementalIndexRow> incrementalIndexRowComparator,
+        List<DimensionDesc> dimensionDescsList
+    )
+    {
+      this.facts = new ConcurrentSkipListMap<>(incrementalIndexRowComparator);
+      this.dimensionDescsList = dimensionDescsList;
+    }
+
+    @Override
+    public int getPriorIndex(IncrementalIndexRow key)
+    {
+      IncrementalIndexRow row = facts.get(key);
+      return row == null ? IncrementalIndexRow.EMPTY_ROW_INDEX : row.getRowIndex();
+    }
+
+    @Override
+    public long getMinTimeMillis()
+    {
+      return facts.firstKey().getTimestamp();
+    }
+
+    @Override
+    public long getMaxTimeMillis()
+    {
+      return facts.lastKey().getTimestamp();
+    }
+
+    @Override
+    public Iterator<IncrementalIndexRow> iterator(boolean descending)
+    {
+      if (descending) {
+        return facts.descendingMap()
+                    .keySet()
+                    .iterator();
+      }
+      return keySet().iterator();
+    }
+
+    @Override
+    public Iterable<IncrementalIndexRow> timeRangeIterable(boolean descending, long timeStart, long timeEnd)
+    {
+      IncrementalIndexRow start = new IncrementalIndexRow(timeStart, new Object[]{}, dimensionDescsList);
+      IncrementalIndexRow end = new IncrementalIndexRow(timeEnd, new Object[]{}, dimensionDescsList);
+      ConcurrentNavigableMap<IncrementalIndexRow, IncrementalIndexRow> subMap = facts.subMap(start, end);
+      ConcurrentMap<IncrementalIndexRow, IncrementalIndexRow> rangeMap = descending ? subMap.descendingMap() : subMap;
+      return rangeMap.keySet();
+    }
+
+    @Override
+    public Iterable<IncrementalIndexRow> keySet()
+    {
+      return facts.keySet();
+    }
+
+    @Override
+    public Iterable<IncrementalIndexRow> persistIterable()
+    {
+      // with rollup, facts are already pre-sorted so just return keyset
+      return keySet();
+    }
+
+    @Override
+    public int putIfAbsent(IncrementalIndexRow key, int rowIndex)
+    {
+      // setRowIndex() must be called before facts.putIfAbsent() for visibility of rowIndex from concurrent readers.
+      key.setRowIndex(rowIndex);
+      IncrementalIndexRow prev = facts.putIfAbsent(key, key);
+      return prev == null ? IncrementalIndexRow.EMPTY_ROW_INDEX : prev.getRowIndex();
+    }
+
+    @Override
+    public void clear()
+    {
+      facts.clear();
+    }
+  }
+
+  static final class PlainFactsHolder implements FactsHolder
+  {
+    private final ConcurrentNavigableMap<Long, Deque<IncrementalIndexRow>> facts;
+
+    private final Comparator<IncrementalIndexRow> incrementalIndexRowComparator;
+
+    public PlainFactsHolder(Comparator<IncrementalIndexRow> incrementalIndexRowComparator)
+    {
+      this.facts = new ConcurrentSkipListMap<>();
+      this.incrementalIndexRowComparator = incrementalIndexRowComparator;
+    }
+
+    @Override
+    public int getPriorIndex(IncrementalIndexRow key)
+    {
+      // always return EMPTY_ROW_INDEX to indicate that no prior key cause we always add new row
+      return IncrementalIndexRow.EMPTY_ROW_INDEX;
+    }
+
+    @Override
+    public long getMinTimeMillis()
+    {
+      return facts.firstKey();
+    }
+
+    @Override
+    public long getMaxTimeMillis()
+    {
+      return facts.lastKey();
+    }
+
+    @Override
+    public Iterator<IncrementalIndexRow> iterator(boolean descending)
+    {
+      if (descending) {
+        return timeOrderedConcat(facts.descendingMap().values(), true).iterator();
+      }
+      return timeOrderedConcat(facts.values(), false).iterator();
+    }
+
+    @Override
+    public Iterable<IncrementalIndexRow> timeRangeIterable(boolean descending, long timeStart, long timeEnd)
+    {
+      ConcurrentNavigableMap<Long, Deque<IncrementalIndexRow>> subMap = facts.subMap(timeStart, timeEnd);
+      final ConcurrentMap<Long, Deque<IncrementalIndexRow>> rangeMap = descending ? subMap.descendingMap() : subMap;
+      return timeOrderedConcat(rangeMap.values(), descending);
+    }
+
+    private Iterable<IncrementalIndexRow> timeOrderedConcat(
+        final Iterable<Deque<IncrementalIndexRow>> iterable,
+        final boolean descending
+    )
+    {
+      return () -> Iterators.concat(
+          Iterators.transform(
+              iterable.iterator(),
+              input -> descending ? input.descendingIterator() : input.iterator()
+          )
+      );
+    }
+
+    private Stream<IncrementalIndexRow> timeAndDimsOrderedConcat(
+        final Collection<Deque<IncrementalIndexRow>> rowGroups
+    )
+    {
+      return rowGroups.stream()
+                      .flatMap(Collection::stream)
+                      .sorted(incrementalIndexRowComparator);
+    }
+
+    @Override
+    public Iterable<IncrementalIndexRow> keySet()
+    {
+      return timeOrderedConcat(facts.values(), false);
+    }
+
+    @Override
+    public Iterable<IncrementalIndexRow> persistIterable()
+    {
+      return () -> timeAndDimsOrderedConcat(facts.values()).iterator();
+    }
+
+    @Override
+    public int putIfAbsent(IncrementalIndexRow key, int rowIndex)
+    {
+      Long time = key.getTimestamp();
+      Deque<IncrementalIndexRow> rows = facts.get(time);
+      if (rows == null) {
+        facts.putIfAbsent(time, new ConcurrentLinkedDeque<>());
+        // in race condition, rows may be put by other thread, so always get latest status from facts
+        rows = facts.get(time);
+      }
+      // setRowIndex() must be called before rows.add() for visibility of rowIndex from concurrent readers.
+      key.setRowIndex(rowIndex);
+      rows.add(key);
+      // always return EMPTY_ROW_INDEX to indicate that we always add new row
+      return IncrementalIndexRow.EMPTY_ROW_INDEX;
+    }
+
+    @Override
+    public void clear()
+    {
+      facts.clear();
     }
   }
 }
