+++ processing/src/test/java/org/apache/druid/query/aggregation/mean/SimpleTestIndex.java
@@ -101,7 +101,6 @@ public class SimpleTestIndex
           },
           0,
           Granularities.NONE,
-          false,
           100,
           false
       );
