+++ processing/src/test/java/org/apache/druid/query/groupby/GroupByMultiSegmentTest.java
@@ -130,7 +130,6 @@ public class GroupByMultiSegmentTest
                 .withRollup(withRollup)
                 .build()
         )
-        .setConcurrentEventAdd(true)
         .setMaxRowCount(1000)
         .build();
   }
